requirements.txt
discord.py==1.7.3
psycopg2==2.9.3
pytz==2021.3
yoyo-migrations==7.3.2
